mysql query iterator complete
The iterator is complete, which returns a nested object that contains
all the materials and resources needed for a blueprint.
Seems to work with any t1 and t2 blueprint except
for t3 blueprints (for obviours reasons). The next step is to implement
a way to fetch all the resources in the object into a handy list.

Now we can work on a