Changes on how data is proccessed after query
First the 'result' object is created from repeated sql queries
This object is then processed in the function 'processObj', this returns
an object called 'processedObj` where the index is the itemID and the
property is the total number of this item required for a single
blueprint run for the queried blueprint, in this case for the
'retribution' ship.

processedObj can be used to display useful information to the user.